+++ NWN.Anvil/src/main/API/Events/Game/PlaceableEvents/PlaceableEvents.OnClose.cs
@@ -1,0 +1,47 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwPlaceable"/>.
+  /// </summary>
+  public static partial class PlaceableEvents
+  {
+    /// <summary>
+    /// Called when <see cref="NwCreature"/> has closed a <see cref="NwPlaceable"/>.
+    /// </summary>
+    [GameEvent(EventScriptType.PlaceableOnClosed)]
+    public sealed class OnClose : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwPlaceable"/> that was closed.
+      /// </summary>
+      public NwPlaceable Placeable { get; } = NWScript.OBJECT_SELF.ToNwObject<NwPlaceable>();
+
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that closed the <see cref="NwPlaceable"/>.
+      /// </summary>
+      public NwCreature ClosedBy { get; } = NWScript.GetLastClosedBy().ToNwObject<NwCreature>();
+
+      NwObject IEvent.Context
+      {
+        get => Placeable;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwPlaceable
+  {
+    /// <inheritdoc cref="PlaceableEvents.OnClose"/>
+    public event Action<PlaceableEvents.OnClose> OnClose
+    {
+      add => EventService.Subscribe<PlaceableEvents.OnClose, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<PlaceableEvents.OnClose, GameEventFactory>(this, value);
+    }
+  }
+}
