+++ NWN.Anvil/src/main/API/Events/Game/PlaceableEvents/PlaceableEvents.OnDisturbed.cs
@@ -1,0 +1,57 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwPlaceable"/>.
+  /// </summary>
+  public static partial class PlaceableEvents
+  {
+    /// <summary>
+    /// Called when <see cref="NwPlaceable"/> inventory has been disturbed.
+    /// </summary>
+    [GameEvent(EventScriptType.PlaceableOnInventoryDisturbed)]
+    public sealed class OnDisturbed : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="InventoryDisturbType"/>.
+      /// </summary>
+      public InventoryDisturbType DisturbType { get; } = (InventoryDisturbType)NWScript.GetInventoryDisturbType();
+
+      /// <summary>
+      /// Gets the <see cref="NwPlaceable"/> that was disturbed.
+      /// </summary>
+      public NwPlaceable Placeable { get; } = NWScript.OBJECT_SELF.ToNwObject<NwPlaceable>();
+
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that disturbed <see cref="NwPlaceable"/>.
+      /// </summary>
+      public NwCreature Disturber { get; } = NWScript.GetLastDisturbed().ToNwObject<NwCreature>();
+
+      /// <summary>
+      /// Gets the <see cref="NwItem"/> that triggered the disturb event on <see cref="NwPlaceable"/>.
+      /// </summary>
+      public NwItem DisturbedItem { get; } = NWScript.GetInventoryDisturbItem().ToNwObject<NwItem>();
+
+      NwObject IEvent.Context
+      {
+        get => Placeable;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwPlaceable
+  {
+    /// <inheritdoc cref="PlaceableEvents.OnDisturbed"/>
+    public event Action<PlaceableEvents.OnDisturbed> OnDisturbed
+    {
+      add => EventService.Subscribe<PlaceableEvents.OnDisturbed, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<PlaceableEvents.OnDisturbed, GameEventFactory>(this, value);
+    }
+  }
+}
