+++ NWN.Anvil/src/main/API/Events/Game/PlaceableEvents/PlaceableEvents.OnHeartbeat.cs
@@ -1,0 +1,42 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwPlaceable"/>.
+  /// </summary>
+  public static partial class PlaceableEvents
+  {
+    /// <summary>
+    /// Called every 6 seconds on <see cref="NwPlaceable"/>.
+    /// </summary>
+    [GameEvent(EventScriptType.PlaceableOnHeartbeat)]
+    public sealed class OnHeartbeat : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwPlaceable"/> heartbeat.
+      /// </summary>
+      public NwPlaceable Placeable { get; } = NWScript.OBJECT_SELF.ToNwObject<NwPlaceable>();
+
+      NwObject IEvent.Context
+      {
+        get => Placeable;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwPlaceable
+  {
+    /// <inheritdoc cref="PlaceableEvents.OnHeartbeat"/>
+    public event Action<PlaceableEvents.OnHeartbeat> OnHeartbeat
+    {
+      add => EventService.Subscribe<PlaceableEvents.OnHeartbeat, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<PlaceableEvents.OnHeartbeat, GameEventFactory>(this, value);
+    }
+  }
+}
