+++ NWN.Anvil/src/main/API/Events/Game/PlaceableEvents/PlaceableEvents.OnPhysicalAttacked.cs
@@ -1,0 +1,68 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwPlaceable"/>.
+  /// </summary>
+  public static partial class PlaceableEvents
+  {
+    /// <summary>
+    /// Called when <see cref="NwPlaceable"/> has been physically attacked.
+    /// </summary>
+    [GameEvent(EventScriptType.PlaceableOnMeleeAttacked)]
+    public sealed class OnPhysicalAttacked : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwPlaceable"/> that was physically attacked.
+      /// </summary>
+      public NwPlaceable Placeable { get; } = NWScript.OBJECT_SELF.ToNwObject<NwPlaceable>();
+
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that attacked the <see cref="NwPlaceable"/>.
+      /// </summary>
+      public NwCreature Attacker { get; } = NWScript.GetLastAttacker().ToNwObject<NwCreature>();
+
+      /// <summary>
+      /// Gets the <see cref="NwItem"/> used to damage <see cref="NwPlaceable"/>.
+      /// </summary>
+      public NwItem WeaponUsed(NwCreature attacker)
+      {
+        return NWScript.GetLastWeaponUsed(attacker).ToNwObject<NwItem>();
+      }
+
+      /// <summary>
+      /// Gets the <see cref="SpecialAttack"/> used to damage <see cref="NwPlaceable"/>.
+      /// </summary>
+      public SpecialAttack AttackType { get; } = (SpecialAttack)NWScript.GetLastAttackType();
+
+      /// <summary>
+      /// Gets the <see cref="ActionMode"/> used to damage <see cref="NwPlaceable"/>.
+      /// </summary>
+      public ActionMode AttackMode(NwCreature attacker)
+      {
+        return (ActionMode)NWScript.GetLastAttackMode(attacker);
+      }
+
+      NwObject IEvent.Context
+      {
+        get => Placeable;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwPlaceable
+  {
+    /// <inheritdoc cref="PlaceableEvents.OnPhysicalAttacked"/>
+    public event Action<PlaceableEvents.OnPhysicalAttacked> OnPhysicalAttacked
+    {
+      add => EventService.Subscribe<PlaceableEvents.OnPhysicalAttacked, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<PlaceableEvents.OnPhysicalAttacked, GameEventFactory>(this, value);
+    }
+  }
+}
